chore(apps/api): create separate controller for GetFuseLeveragedTokensGains
