+++ apps/api/src/controllers/flt/index.ts
@@ -2,12 +2,14 @@ import GetFuseLeveragedTokensByChainId from "./GetFuseLeveragedTokensByChainId";
 import GetFuseLeveragedTokenBySymbol from "./GetFuseLeveragedTokenBySymbol";
 import GetFuseLeveragedTokenUserPositionById from "./GetFuseLeveragedTokenUserPositionById";
 import GetFuseLeveragedTokenSwapsBySymbol from "./GetFuseLeveragedTokenSwapsBySymbol";
+import GetFuseLeveragedTokensGains from "./GetFuseLeveragedTokensGains";
 
 const controllers = {
     GetFuseLeveragedTokensByChainId,
     GetFuseLeveragedTokenBySymbol,
     GetFuseLeveragedTokenUserPositionById,
     GetFuseLeveragedTokenSwapsBySymbol,
+    GetFuseLeveragedTokensGains,
 };
 
 export default controllers;
